kernel-netlink: Read last use time from SA if possible
Since 5.17, the Linux kernel updates the last use time per SA.  In
previous releases the attribute was only updated and reported for
specific outbound IPv6 SAs.

Using this reduces the number of kernel queries per CHILD_SA:  for DPDs
from two policy queries (IN/FWD) to a single query of the inbound SA and
for status reports, the three policy queries (IN/FWD/OUT) can be omitted
and only the two SAs have to be queried.  For NAT keepalives the number
of queries doesn't change, instead a policy query (OUT) is replaced by
a query for the outbound SA.

While we could use the existence of the attribute as indicator for its
support, we don't know this until we queried an SA.  By using a version
check we can announce the feature from the start.